requirements.txt
streamlit==1.24.0
openai==1.0.0
pandas==2.0.3
python-dotenv==1.1.0
httpx==0.24.1